Towards immutable everything
This touches on the main problem with tui, as far as I see it.

When widgets update the buffer for a given coordinate, it typically patches what is already there. This means that often styles are drawn first, and then text (or vice versa). Especially, widgets can clobber each other's output in this manner, if they are drawn in a hierarchy.

We'll do this inheritance of styles explicitly instead.

This commit sketches an API for composing styles, so you can easily compose things like  `style / header_style / label`, with the equivalent result of applying the two styles and one content after each other with the old API.